Rename file to typescript
[MAILPOET-2835]
