+++ assets/js/src/form_editor/blocks/divider/edit.tsx
